REFACTOR: javascript floating point calculations sometimes result in pricing having more than 2 decimals. Refactors sequelize tables to use integers and price values to be in cents

+++ server/db/models/product.js
@@ -17,7 +17,7 @@ const Product = db.define("product", {
     }
   },
   price: {
-    type: Sequelize.FLOAT
+    type: Sequelize.INTEGER // THESE ARE IN CENTS
   },
   size: {
     type: Sequelize.STRING,
